Split protocol.c into common and dmm modules
